Move data explorer constants to new file, refactor and rename constants
