+++ app/javascript/app/components/data-explorer-content/data-explorer-content.js
@@ -12,15 +12,15 @@ import {
   DATA_EXPLORER_FILTERS,
   DATA_EXPLORER_EXTERNAL_PREFIX,
   DATA_EXPLORER_DEPENDENCIES,
-  DATA_EXPLORER_PER_PAGE,
-  ESP_HOST
-} from 'data/constants';
+  DATA_EXPLORER_PER_PAGE
+} from 'data/data-explorer-constants';
+import { ESP_HOST } from 'data/constants';
 import DataExplorerContentComponent from './data-explorer-content-component';
 import {
   parseData,
   getMethodology,
-  getFirstColumnHeaders,
   getSectionLabel,
+  getFirstTableHeaders,
   getFilteredOptions,
   getSelectedOptions,
   getPathwaysMetodology,
@@ -51,7 +51,7 @@ const mapStateToProps = (state, { section, location }) => {
   const devESPURL = section === 'emission-pathways' ? ESP_HOST : '';
   const downloadHref = `${devESPURL}/api/v1/data/${DATA_EXPLORER_SECTIONS[
     section
-  ].requestPath}/download.csv${filterQuery ? `?${filterQuery}` : ''}`;
+  ].label}/download.csv${filterQuery ? `?${filterQuery}` : ''}`;
   const meta =
     section === 'emission-pathways'
       ? getPathwaysMetodology(dataState)
@@ -89,7 +89,7 @@ const mapStateToProps = (state, { section, location }) => {
     meta,
     metadataSection: !!location.hash && location.hash === '#meta',
     isDisabled,
-    firstColumnHeaders: getFirstColumnHeaders(dataState),
+    firstColumnHeaders: getFirstTableHeaders(dataState),
     href: getLink(dataState),
     sectionLabel: getSectionLabel(dataState),
     downloadHref,
